flask/requirements.txt
appdirs==1.4.4
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2022.6.15
chardet==4.0.0
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.5
cssselect==1.1.0
fake-useragent==0.1.11
Flask==2.1.2
googlesearch-python==1.1.0
gunicorn==20.1.0
idna==2.10
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.0
MarkupSafe==2.1.1
numpy==1.23.0
pandas==1.4.3
parse==1.19.0
pyee==8.2.2
pyppeteer==1.0.2
pyquery==1.4.3
python-dateutil==2.8.2
pytz==2022.1
requests==2.25.1
requests-html==0.10.0
six==1.16.0
soupsieve==2.3.2.post1
tqdm==4.64.0
urllib3==1.26.9
w3lib==1.22.0
websockets==10.3
Werkzeug==2.1.2
zipp==3.8.0
